Validation[COM-3]

+++ src/main/java/com/example/registration/servletAllMagaz/DeleteMagazineServlet.java
@@ -18,17 +18,36 @@ public class DeleteMagazineServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String delete = request.getParameter("delete");
+        Connection con = null;
+        Statement pst = null;
+        if (delete != null) {
+            try {
+                con = DBCPDataSource.getConnection();
+                pst = con.createStatement();
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
 
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            Connection con = DBCPDataSource.getConnection();
-            Statement pst = con.createStatement();
-            String sql = "delete from mydb.magazines where id= '"+delete+"'";
-            pst.executeUpdate(sql);
+            try {
+                Class.forName("com.mysql.jdbc.Driver");
 
-        } catch (ClassNotFoundException | SQLException e) {
-            throw new RuntimeException(e);
+                String sql = "delete from mydb.magazines where id= '" + delete + "'";
+                pst.executeUpdate(sql);
+
+            } catch (ClassNotFoundException | SQLException e) {
+                throw new RuntimeException(e);
+            } finally {
+                try {
+                    pst.close();
+                    con.close();
+                } catch (SQLException e) {
+                    throw new RuntimeException(e);
+                }
+            }
+            response.sendRedirect("store.jsp");
+        } else {
+            response.sendRedirect("error_data.jsp");
         }
-        response.sendRedirect("store.jsp");
+
     }
 }
